Refactor sessionx._
sessionx was really difficult to understand with the layout and
seemingly random location for different types and functions.

This commit specifically addresses this by presenting a more easily
understood way: a file per concept. Other files were touched to deal
with differing imports.

Further, this commit does away with extraneous implicits and reworks
some of the implicits using a common and modern typeclass pattern

Fixes: #39
