+++ core/src/main/scala/com/lookout/borderpatrol/sessionx/SecretStores.scala
@@ -1,0 +1,50 @@
+package com.lookout.borderpatrol.sessionx
+
+/**
+ * Two secrets must be in rotation at any given time:
+ *  - Current: used for creating new sessions and validating incoming non-expired sessions
+ *  - Previous: validating incoming non-expired sessions, e.g. sessions signed by yesterday's key
+ *
+ * Since each [[com.lookout.borderpatrol.sessionx.Secret Secret]] expires (default of 1 day), the
+ * window of a non-expired [[com.lookout.borderpatrol.sessionx.Session Session]] is somewhere between the expiry of
+ * the current `Secret` and the previous `Secret`
+ */
+trait SecretStoreApi {
+  def current: Secret
+  def previous: Secret
+  def find(f: Secret => Boolean): Option[Secret]
+}
+
+/**
+ * Default implementations of [[com.lookout.borderpatrol.sessionx.SecretStoreApi SecretStoreApi]]
+ */
+object SecretStores {
+
+  /**
+   * A useful [[com.lookout.borderpatrol.sessionx.Secrets Secrets]] mock store for quickly testing and prototyping
+   *
+   * @param secrets the current secret and previous secret
+   */
+  case class InMemorySecretStore(secrets: Secrets) extends SecretStoreApi {
+    @SuppressWarnings(Array("org.brianmckenna.wartremover.warts.Var")) // this is for mocking
+    private[this] var _secrets: Secrets = secrets
+
+    def current: Secret = {
+      val c = _secrets.current
+      if (c.expired) {
+        _secrets = _secrets.copy(Secret(), c)
+        _secrets.current
+      }
+      else c
+    }
+
+    def previous: Secret =
+      _secrets.previous
+
+    def find(f: (Secret) => Boolean): Option[Secret] =
+      if (f(current)) Some(current)
+      else if (f(previous)) Some(previous)
+      else None
+  }
+
+}
