+++ server/src/main/scala/com/lookout/borderpatrol/routers/ServiceMatchers.scala
@@ -5,6 +5,38 @@ import com.twitter.finagle.httpx.path.Path
 import com.twitter.util.{Try, Future}
 import io.finch.route._
 
+/**
+ * This provides the Finch compatible [[io.finch.route.Router]] types for composing the logic behind Border Patrol's
+ * service inference.
+ *
+ * We derive a service `name` (a [[String]] name referencing a [[com.twitter.finagle.Name Name]]) either via the `Path`
+ * or given a path of `/` we use the subdomain of the host.
+ *
+ * There are two ways to determine the intending service to route to:
+ *  1) `http://example.com/:service/rest/of/path` => extract `:service`, then lookup a matching `ServiceIdentifier`
+ *  2) `http://service.example.com/` => if path of `/`, extract subdomain `service`, lookup a matching
+ *  `ServiceIdentifier`
+ *
+ * `Path` based matching must _always_ override the fallback `subdomain` matching.
+ *
+ * @example
+ *          Let's say we have a service named "enterprise", if we
+ *          define the instance of that as:
+ *            {{{
+ *            val ent = Service("enterprise", Path("/ent"), "default")
+ *            val biz = Service("business", Path("/biz"), "api")
+ *            }}}
+ *
+ *          The following urls would match the services:
+ *            - api.example.com/ent => ent
+ *            - default.example.com/ => ent
+ *            - default.example.com/biz => biz
+ *            - api.example.com => biz
+ *            - api.example.com/ent => ent
+ *          These would not match the services:
+ *            - a.api.example.com => ???
+ *            - example.com => ???
+ */
 object ServiceMatchers {
 
   /**
@@ -32,23 +64,30 @@ object ServiceMatchers {
     val domainTerm = "."
     val pathTerm = "/"
 
+    // helper for finding longest subdomain prefix in a set
+    private[this] def longestPrefix(sis: Set[ServiceIdentifier]): Option[ServiceIdentifier] =
+      sis.foldRight(Option.empty[ServiceIdentifier])((si, res) => res match {
+        case Some(s) if si.subdomain.size < s.subdomain.size => Some(s)
+        case _ => Some(si)
+      })
+
     /**
-     * Find the longest matching subdomain that matches the host
+     * Gives the name of the service that best matches the subdomain of the host string, or None
      *
      * @example
      *          Given a host of "sub.subdomain.example.org" and a Set[ServiceIdentifier] of
-     *            (ServiceIdentifier("one", "/s", "sub.subdomain"),
-     *             ServiceIdentifier("two", "/s2", "sub"))
-   *             return the ServiceIdentifier named "one" because it is the longest matching
+     *          {{{
+     *            Set(ServiceIdentifier("one", "/s", "sub.subdomain"),
+     *                ServiceIdentifier("two", "/s2", "sub"))
+     *          }}}
+     *          return the [[ServiceIdentifier.name]] "one" because it is the longest matching
      * @param host The fully qualified host name
      * @return the service name from the longest matching subdomain
      */
     def subdomain(host: String)(implicit services: Set[ServiceIdentifier]): Option[String] =
-      services.filter(si => host.startsWith(si.subdomain + domainTerm))
-              .foldRight(Option.empty[ServiceIdentifier])((si, res) => res match {
-                case Some(s) if si.subdomain.size < s.subdomain.size => Some(s)
-                case _ => Some(si)})
-              .map(_.name)
+      longestPrefix(
+        services.filter(si => host.startsWith(si.subdomain + domainTerm))
+      ).map(_.name)
 
     /**
      * Find the path exactly matching the path in the request
